Add From with f64 arrays for Color

+++ src/os/macos/view.rs
@@ -52,10 +52,10 @@ impl View {
     pub(crate) fn set_background_color(&mut self, color: Color) {
         unsafe {
             let color: id = msg_send![class("NSColor"), 
-                colorWithRed: color.inner.red
-                       green: color.inner.green
-                        blue: color.inner.blue
-                       alpha: color.inner.alpha];
+                colorWithRed: color.inner.red as f64
+                       green: color.inner.green as f64
+                        blue: color.inner.blue as f64
+                       alpha: color.inner.alpha as f64];
 
             msg_send![color, retain];
